requirements.txt
Flask
gunicorn
pytest
flask-restful
python-decouple
pandas
click~=7.1.2
gspread
